requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
pydantic-settings
python-dotenv
celery
redis
requests
beautifulsoup4
sentence-transformers
torch
pandas
streamlit
litellm
numpy
scipy
playwright
openpyxl
